only valid/invalid
